requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-jazzmin==2.6.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
yarl==1.9.4
